requirements.txt
streamlit
python-dotenv
langchain
langchain_groq
langchain-openai
faiss-cpu
langchain-community
